+++ lib/salesforce_ar_sync/jobs/delete_object_job.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module SalesforceArSync
+  class DeleteObjectJob < BaseJob
+    def perform(klass, sobject)
+      prepare_class(klass).delete_object(sobject)
+    end
+  end
+end