+++ lib/salesforce_ar_sync/jobs/sync_object_job.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module SalesforceArSync
+  class SyncObjectJob < BaseJob
+    def perform(klass, sobject)
+      prepare_class(klass).salesforce_update(sobject)
+    end
+  end
+end