+++ lib/salesforce_ar_sync/soap_handler/base.rb
@@ -13,7 +13,10 @@ module SalesforceArSync
       # queues each individual record from the message for update
       def process_notifications(priority = 90)
         batch_process do |sobject|
-          options[:klass].camelize.constantize.delay(priority: priority, run_at: 5.seconds.from_now).salesforce_update(sobject)
+          SalesforceArSync::SyncObjectJob.set(
+            priority: priority,
+            wait_until: 5.seconds.from_now
+          ).perform_later(options[:klass], sobject)
         end
       end